fix(gateway): M4 decimal wallet projection + M5 remove dead pricing override (JEBV4-49) (#282)
M4 — money is now decimal end-to-end in the wallet READ projection. The ledger
reader cast the NUMERIC column to double ((double)reader.GetDecimal(2)), and the
balance/ledger DTOs were double — large LBP amounts past 2^53 lose integer
precision and fractional artifacts leak into a money display. JeebWalletResponses
(AvailableBalance/ReservedNow/GiftCredit/LedgerEntry.Amount) and the projection
threshold are now decimal; the ledger reader reads GetDecimal straight through
(no double cast). The generic wallet-service primitive Amount stays double (NSwag
reusable-client boundary) so ProjectBalance converts ONCE at that boundary and
keeps decimal thereafter.

M5 — deleted the dead JeebPricingOptions "config-overridable rates" affordance
and its Program.cs registration. SettlementService calls the static
CommissionCalculator directly, so a JeebPricing appsettings override never reached
settlement math (a false affordance / quote-vs-settlement divergence risk).
CommissionCalculator's flat-10% constants are the single source of truth;
JEBV4-43's CommissionAgreementTests guards that policy (green after the delete).

M6 (refund-FX currency) is owner-gated — no code; see the Jira decision comment
(the described USD-vs-LBP mismatch appears resolved: settlement is now USD on main).

Tests: JeebWalletProjectionTests adds decimal-precision coverage (>2^53 integer
preserved, fractional value clean in JSON); Commission* suites pass unchanged.

+++ src/JeebGateway/JeebWallet/JeebWalletLedgerReader.cs
@@ -107,7 +107,11 @@ public sealed class PostgresJeebWalletLedgerReader : IJeebWalletLedgerReader
                 {
                     Id = reader.GetString(0),
                     Type = reader.GetString(1),
-                    Amount = (double)reader.GetDecimal(2),
+                    // JEBV4-49 (M4): keep money as decimal end-to-end — read the
+                    // NUMERIC column straight into the decimal DTO, no (double) cast
+                    // (a double cast can lose integer precision past 2^53 and
+                    // reintroduce fractional artifacts on large LBP amounts).
+                    Amount = reader.GetDecimal(2),
                     Sign = reader.GetInt32(3),
                     Ref = reader.GetString(4),
                     Ts = reader.GetFieldValue<DateTime>(5).ToUniversalTime().ToString("o"),
